# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "backend",
    "description": "12-factor Rust backend: layered store + RBAC + JWT + WS + worker",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/bookings/{id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "`GET /api/admin/bookings/{id}` — admin booking detail.",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Booking ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBookingDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "`PATCH /api/admin/bookings/{id}` — update booking status (admin override).",
        "operationId": "update_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Booking ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBookingStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/admin/payments": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "`GET /api/admin/payments` — paginated payment list.",
        "operationId": "list_admin_payments",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPaymentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/payments/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "`PATCH /api/admin/payments/{id}` — admin override of payment status.",
        "description": "Used to manually mark a payment as `completed` (when the IPN failed\nbut the admin confirmed via the provider's dashboard) or `refunded`\n(when a refund was issued out-of-band).",
        "operationId": "update_payment_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Payment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentStatusReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Payment status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePaymentStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/admin/routes": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "`GET /api/admin/routes` — list routes (admin), with optional brand\nfilter, search and offset pagination.",
        "operationId": "list",
        "parameters": [
          {
            "name": "brandId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive search over the route name, start/end city\nslugs and the owning brand's name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (clamped to `[1, 200]` by the service). `None` = all\nrows (legacy consumers).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Page offset (0-based).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRouteListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "`POST /api/admin/routes` — create a route.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/routes/{id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "`PUT /api/admin/routes/{id}` — update a route.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Route ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "`DELETE /api/admin/routes/{id}` — delete a route.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Route ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/bookings/{id}/confirm": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "`POST /api/bookings/{id}/confirm` — confirm a booking with payment.",
        "operationId": "confirm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Booking ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Booking confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfirmResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/payments": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "`POST /api/payments` — initiate a payment for a booking.",
        "operationId": "create_payment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Payment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePaymentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — not your booking"
          },
          "404": {
            "description": "Booking not found"
          },
          "409": {
            "description": "Booking not in pending status"
          },
          "503": {
            "description": "Provider not configured"
          }
        }
      }
    },
    "/api/payments/booking/{bookingId}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "`GET /api/payments/booking/{bookingId}` — list payments for a booking.",
        "operationId": "list_booking_payments",
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "description": "Booking ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPaymentsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — not your booking"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
    "/api/payments/ipn/momo": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "`POST /api/payments/ipn/momo` — MoMo IPN webhook.",
        "description": "MoMo POSTs JSON to `ipnUrl`. We verify the HMAC-SHA256 signature\nover the canonical alphabetical-key string. Response: HTTP 204\nwith body `{\"code\":0,\"message\":\"...\"}`.",
        "operationId": "momo_ipn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "IPN processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpnResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body"
          },
          "401": {
            "description": "Invalid signature"
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    },
    "/api/payments/ipn/vnpay": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "`GET /api/payments/ipn/vnpay` — VNPay IPN webhook.",
        "description": "VNPay POSTs the IPN as a GET request (with all params in the URL).\nWe re-verify the HMAC-SHA512 signature and transition the payment state.\nResponse body must be JSON `{\"RspCode\":\"00\",\"Message\":\"Confirm Success\"}`.",
        "operationId": "vnpay_ipn",
        "parameters": [
          {
            "name": "params",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              },
              "propertyNames": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "IPN processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpnResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid params"
          },
          "401": {
            "description": "Invalid signature"
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    },
...
    "/api/payments/{id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "`GET /api/payments/{id}` — get payment status (for polling).",
        "operationId": "get_payment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Payment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — not your payment"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminBookingDetailResponse": {
        "type": "object",
        "description": "Response of `GET /api/admin/bookings/{id}`.",
        "required": [
          "item"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/AdminBookingDetail"
          }
        }
      },
...
      "UpdateBookingStatusRequest": {
        "type": "object",
        "description": "Request body for `PATCH /api/admin/bookings/{id}`.",
        "required": [
          "status"
        ],
        "properties": {
          "force": {
            "type": "boolean",
            "description": "When `true`, skips the state-machine transition check (admin override)."
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "description": "`pending` | `confirmed` | `paid` | `completed` | `cancelled` | `refunded`.\n`paid` is normalized to `confirmed`; `refunded` to `cancelled`."
          }
        }
      },
      "UpdateBookingStatusResponse": {
        "type": "object",
        "description": "Response of `PATCH /api/admin/bookings/{id}`.",
        "required": [
          "item"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/AdminBookingStatusUpdate"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}